env: add config for minify, IE11
