--- apps/react-calendar/src/ie11.ts
@@ -1,7 +1,0 @@
-// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-// @ts-ignore
-import ToastUICalendar from '@toast-ui/calendar/ie11';
-
-import { createComponent } from './calendar';
-
-export default createComponent(ToastUICalendar);
