--- apps/react-calendar/src/index.ts
@@ -1,5 +1,0 @@
-import ToastUICalendar from '@toast-ui/calendar';
-
-import { createComponent } from './calendar';
-
-export default createComponent(ToastUICalendar);
